+++ src/app.routes.ts
@@ -5,7 +5,7 @@ import { Documentation } from './app/pages/documentation/documentation';
 import { Landing } from './app/pages/landing/landing';
 import { Notfound } from './app/pages/notfound/notfound';
 import { authGuard } from './app/guards/auth.guard';
-import {ComponentesPage} from './app/pages/gestion/componentes'
+import {HomeComponent} from './app/pages/home/home.component'
 
 export const appRoutes: Routes = [
     {
@@ -13,8 +13,8 @@ export const appRoutes: Routes = [
         component: AppLayout,
         canActivate: [authGuard],
         children: [
-            //{ path: '', component: Dashboard },
-            { path: '', component: ComponentesPage },
+            { path: 'dashboard', component: Dashboard },
+            { path: '', component: HomeComponent },
             { path: 'uikit', loadChildren: () => import('./app/pages/uikit/uikit.routes') },
             { path: 'documentation', component: Documentation },
             { path: 'pages', loadChildren: () => import('./app/pages/pages.routes') }